Phase 0: fork foundation — rename, bug fixes, tests, CI
- Rename module github.com/dece2183/yamusic-tui -> github.com/lillink13/yamusic-tui
- Fix LRC lyrics desync: parse [mm:ss.xx] fraction as centi/milliseconds
  (was treated as raw ms, desyncing up to ~1s); skip [ar:]/[al:] metadata lines
- Propagate previously-swallowed json.Decode errors in api request handling
- Backfill missing config scalar defaults (volume/buffer-size/rewind); fix
  show-all-keys yaml tag with backward-compat for legacy show-all-kyes; align
  README to suppress-errors
- Lint cleanups (unconvert, misspell EV_TRACK_SKIPPED, errors.New); repo is now
  golangci-lint clean
- Add unit tests: api/lyrics_test.go, config/config_test.go
- Add .golangci.yml; new ci.yml (gofmt/vet/test/lint + linux+macos build matrix);
  rewrite release.yml to a matrix (linux amd64/arm64, macOS arm64/amd64 with
  external linker, windows amd64, each +/-nomedia; artifacts yamusic-tui-<os>-<arch>)

+++ stream/bufferedStream.go
@@ -41,7 +41,7 @@ func NewBufferedStream(source io.ReadCloser, totalSize int64) *BufferedStream {
 }
 
 func (h *BufferedStream) Length() int64 {
-	return int64(h.totalSize)
+	return h.totalSize
 }
 
 func (h *BufferedStream) Close() error {
